requirements.txt
yfinance>=0.2.0
pandas>=2.0
numpy>=1.24
requests>=2.28
beautifulsoup4>=4.12
python-telegram-bot>=20.0
pyyaml>=6.0
python-dotenv>=1.0
keyring>=25.0
transformers>=4.30
torch>=2.0
feedparser>=6.0
pytest>=7.0
pytest-asyncio>=0.21
